+++ src/components/inheritance/Tab.jsx
@@ -1,0 +1,30 @@
+import { Tabs } from "flowbite-react";
+import { HiAdjustments, HiClipboardList, HiUserCircle } from "react-icons/hi";
+import { MdDashboard } from "react-icons/md";
+import { Link } from "react-router-dom";
+
+export default function Tab() {
+  return (
+    <div>
+      <ul class="flex flex-wrap text-sm font-medium text-center text-gray-500 border-b border-gray-200 dark:border-gray-700 dark:text-gray-400">
+        <li class="me-2">
+          <Link
+            to={"/inhertance"}
+            aria-current="page"
+            className="inline-block p-4 text-blue-600 bg-gray-100 rounded-t-lg active dark:bg-gray-800 dark:text-blue-500"
+          >
+            Profile
+          </Link>
+        </li>
+        <li class="me-2">
+          <Link
+            to={"/property"}
+            className="inline-block p-4 rounded-t-lg hover:text-gray-600 hover:bg-gray-50 dark:hover:bg-gray-800 dark:hover:text-gray-300"
+          >
+            Dashboard
+          </Link>
+        </li>
+      </ul>
+    </div>
+  );
+}
