+++ src/pages/index.js
@@ -28,6 +28,7 @@ export { default as Home } from "./home/Home";
 
 // Inheritance (상속)
 export { default as Inheritance } from "./inheritance/Inheritance";
+export { default as Property } from "./inheritance/component/Property";
 
 // Transfer (이체)
 export { default as AutoTransfer } from "./transfer/AutoTransfer";
