+++ src/pages/inheritance/component/Contract.jsx
@@ -1,0 +1,7 @@
+export default function Contract() {
+  return (
+    <>
+      <h2>계약서 화면</h2>
+    </>
+  );
+}
